requirements.txt
streamlit==0.69.2
torch==1.6.0
torchvision==0.7.0
Pillow==8.0.0
